* Rename directory name to manager/ -- this is because when you run go build . venvy already exists so it cannot create the executable file without renameing use go mod for deps
