+++ go.mod
@@ -1,0 +1,20 @@
+module github.com/pnegahdar/venvy
+
+require (
+	github.com/BurntSushi/toml v0.3.0
+	github.com/fatih/color v1.6.0
+	github.com/go-playground/locales v0.11.2
+	github.com/go-playground/universal-translator v0.16.0
+	github.com/google/btree v0.0.0-20180813153112-4030bb1f1f0c // indirect
+	github.com/inconshreveable/mousetrap v1.0.0
+	github.com/mattn/go-colorable v0.0.9
+	github.com/mattn/go-isatty v0.0.3
+	github.com/mitchellh/go-homedir v0.0.0-20161203194507-b8bc1bf76747
+	github.com/peterbourgon/diskv v2.0.1+incompatible
+	github.com/sirupsen/logrus v1.0.4
+	github.com/spf13/cobra v0.0.1
+	github.com/spf13/pflag v1.0.0
+	golang.org/x/crypto v0.0.0-20180228161326-91a49db82a88
+	golang.org/x/sys v0.0.0-20180302081741-dd2ff4accc09
+	gopkg.in/go-playground/validator.v9 v9.11.0
+)
